updating

+++ src/core/meal/usecase/delete-meal/delete-meal.usecase.ts
@@ -7,11 +7,17 @@ export class DeleteMealUseCase {
     private mealGateway: IMealGateway,
   ) {}
 
-  async execute(id: string): Promise<void> {
+  async execute(id: string, userId: string): Promise<void> {
     const meal = await this.mealGateway.findMealById(id);
 
     if (!meal) {
       throw new BadRequestException('Meal not found!');
+    }
+
+    if (meal.userId !== userId) {
+      throw new BadRequestException(
+        'You do not have permission to delete this meal!',
+      );
     }
 
     const deleteMeal = await this.mealGateway.deleteMeal(id);
